# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: La Playita de Don Concho - Restaurant Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InventoryItem" }
    post:
      operationId: createInventoryItem
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInventoryRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryItem" }
  /inventory/{id}:
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInventoryRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryItem" }
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - name: force
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "409":
          description: In use by recipes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InventoryInUseError" }
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDetail" }
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDetail" }
  /orders/{id}/items:
    patch:
      operationId: patchOrderItems
      tags: [orders]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                add:
                  type: array
                  items: { $ref: "#/components/schemas/NewOrderItem" }
                remove:
                  type: array
                  items: { type: integer }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderDetail" }
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    InventoryInUseError:
      type: object
      required: [error, usedIn]
      properties:
        error: { type: string }
        usedIn:
          type: array
          items:
            type: object
            required: [recipeId, productId, productName]
            properties:
              recipeId: { type: integer }
              productId: { type: integer }
              productName: { type: string }
    InventoryItem:
      type: object
      required: [id, name, unit, currentQuantity, minimumStock, cost]
      properties:
        id: { type: integer }
        name: { type: string }
        unit: { type: string, enum: [g, kg, ml, l, unit] }
        currentQuantity: { type: number }
        minimumStock: { type: number }
        cost: { type: number }
        supplier: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
    CreateInventoryRequest:
      type: object
      required: [name, unit]
      properties:
        name: { type: string }
        unit: { type: string, enum: [g, kg, ml, l, unit] }
        currentQuantity: { type: number }
        minimumStock: { type: number }
        cost: { type: number }
        supplier: { type: string, nullable: true }
...
    NewOrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        notes: { type: string, nullable: true }
    CreateOrderRequest:
      type: object
      required: [tableId, items]
      properties:
        tableId: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/NewOrderItem" }
    OrderItemDetail:
      type: object
      required: [id, productId, productName, quantity, unitPrice, lineTotal]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        notes: { type: string, nullable: true }
        lineTotal: { type: number }
    OrderDetail:
      type: object
      required: [id, tableId, waiterId, status, total, items, createdAt]
      properties:
        id: { type: integer }
        tableId: { type: integer }
        tableNumber: { type: integer }
        areaName: { type: string }
        waiterId: { type: integer }
        waiterName: { type: string }
        status: { type: string }
        total: { type: number }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItemDetail" }
        createdAt: { type: string, format: date-time }
...